refactor: change outbox save and publish logic
