+++ src/main/java/com/servertech/myboard/like/article/application/event/LikeEventOutboxService.java
@@ -5,6 +5,7 @@ import com.servertech.myboard.like.article.domain.LikeEventOutbox;
 import com.servertech.myboard.like.article.domain.LikeEventOutboxRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -13,9 +14,9 @@ import java.util.List;
 public class LikeEventOutboxService {
 	private final LikeEventOutboxRepository outboxRepository;
 
-	//	@TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
-	public void saveLikeEvent(LikeChange likeChange) {
-		outboxRepository.save(LikeEventOutbox.from(likeChange));
+	@Transactional
+	public LikeEventOutbox saveLikeEvent(LikeChange likeChange) {
+		return outboxRepository.save(LikeEventOutbox.from(likeChange));
 	}
 
 	public List<LikeEventOutbox> findAll() {
